+++ CHANGELOG.md
@@ -11,6 +11,8 @@ For details about compatibility between different releases, see the **Commitment
 
 ### Added
 
+- Support for sending end device uplinks using the CLI (see `ttn-lw-cli simulate application-uplink` command).
+
 ### Changed
 
 ### Deprecated